requirements.txt
tensorflow>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
httpx>=0.24.0
pillow>=9.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
anyio>=3.6.0